requirements.txt
alembic==1.9.2
aniso8601==9.0.1
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
coverage==7.0.5
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Migrate==4.0.2
flask-restx==1.0.5
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pip==22.3.1
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.2.1
python-dotenv==0.21.1
pytz==2022.7.1
setuptools==65.6.3
SQLAlchemy==1.4.46
tomli==2.0.1
Werkzeug==2.2.2
wheel==0.37.1
